feat: refactor CmsBlockHtml component and add CmsElementHtml type
